+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 2.0.1+2
+* fix android M FLAG_GRANT_READ_URI_PERMISSION
+## 2.0.1+1
+* Fixed crash caused by third party plug-in request permissions
+## 2.0.1
+* migrate to androidx
+## 1.3.0
+* roll back support
+## 1.2.3+1
+* request READ_EXTERNAL_STORAGE when the file is not in the app directory
+## 1.2.3
+* Optimize permission request logic on android
+## 1.2.2+2
+* fix crash when requestPermission
+## 1.2.2+1
+* migrate to androidx
+## 1.2.2
+* ask for permission when needed
+## 1.2.1
+* fix somethings
 ## 1.2.0
 * Add custom parameters,"type" android,"uti" iOS
 ## 1.1.1